+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - ✨ Added "Known Sources" to the data panel which may include stars, satellites, and other celestial bodies. These can be used to create tracks corresponding to the currently selected sensor.
  - ✨ Added "File Name(s)" column to Sensors data panel with names of files that include imagery for each sensor.
  - 🛠️ Improved consistency of when buttons are enabled / disabled in data panels.
+ - 🛠️ Updated tracks rendering to completely avoid non-visible tracks to improve performance.
  - 🔧 Fixed bug with lasso'ing features where only the last lasso'ed feature would get selected.
 
 ## [1.12.4] - 2026-8-13